+++ src/Core/Baskets/Validators/BasketValidator.php
@@ -11,8 +11,18 @@ class BasketValidator
         return $unique->count() == count($parameters);
     }
 
-    public function inStock($value, $variant, $basketId, $validator)
+    public function inStock($attribute, $value, $parameters, $validator)
     {
-        return app('api')->productVariants()->canAddToBasket($variant['id'], $variant['quantity'] ?? null);
+        return app('api')->productVariants()->canAddToBasket($parameters[0] ?? null, $value);
+    }
+
+    public function minQuantity($attribute, $value, $parameters, $validator)
+    {
+        return $value >= ($parameters[0] ?? 1);
+    }
+
+    public function minBatch($attribute, $value, $parameters, $validator)
+    {
+        return ($value % $parameters[0] ?? 1) === 0;
     }
 }