+++ src/Http/Requests/Baskets/CreateRequest.php
@@ -2,6 +2,7 @@
 
 namespace GetCandy\Api\Http\Requests\Baskets;
 
+use DB;
 use GetCandy\Api\Http\Requests\FormRequest;
 
 class CreateRequest extends FormRequest
@@ -26,11 +27,22 @@ class CreateRequest extends FormRequest
     {
         $rules = [
             'variants' => 'array|unique_lines',
-            'variants.*' => 'in_stock',
-            'variants.*.id' => 'required|hashid_is_valid:product_variants',
-            'variants.*.quantity' => 'required|numeric|min:1|max:10000',
             'basket_id' => 'hashid_is_valid:baskets',
         ];
+
+        $variants = app('api')->productVariants()->getByHashedIds(
+            collect($this->variants)->pluck('id')->toArray()
+        );
+
+        foreach($this->variants as $i => $v) {
+            $variant = $variants->first(function ($variant) use ($v) {
+                return $variant->encodedId() === $v['id'] ?? null;
+            });
+            if ($variant) {
+                $rules["variants.{$i}.quantity"] = 'required|numeric|min:1|max:10000|min_quantity:' . $variant->min_qty . '|min_batch:' . $variant->min_batch . '|in_stock:' . $v['id'] ?? '0';
+            }
+            $rules["variants.{$i}.id"] = 'required|hashid_is_valid:product_variants';
+        }
 
         return $rules;
     }
@@ -38,7 +50,10 @@ class CreateRequest extends FormRequest
     public function messages()
     {
         return [
+            'variants.*.id.hashid_is_valid' => trans('getcandy::validation.hashid_is_valid'),
             'variants.*.quantity.max' => trans('getcandy::validation.max_qty'),
+            'variants.*.quantity.min_quantity' => trans('getcandy::validation.min_qty'),
+            'variants.*.quantity.min_batch' => trans('getcandy::validation.min_batch'),
         ];
     }
 }