+++ src/Http/Resources/Products/ProductTierResource.php
@@ -3,6 +3,7 @@
 namespace GetCandy\Api\Http\Resources\Products;
 
 use GetCandy\Api\Http\Resources\AbstractResource;
+use GetCandy\Api\Http\Resources\Customers\CustomerGroupResource;
 
 class ProductTierResource extends AbstractResource
 {
@@ -19,6 +20,7 @@ class ProductTierResource extends AbstractResource
     public function includes()
     {
         return [
+            'group' => new CustomerGroupResource($this->whenLoaded('group')),
         ];
     }
 }