+++ src/Http/Resources/Products/ProductVariantResource.php
@@ -4,6 +4,7 @@ namespace GetCandy\Api\Http\Resources\Products;
 
 use GetCandy\Api\Http\Resources\AbstractResource;
 use GetCandy\Api\Http\Resources\Taxes\TaxResource;
+use GetCandy\Api\Http\Resources\Assets\AssetResource;
 use GetCandy\Api\Core\Products\Factories\ProductVariantFactory;
 
 class ProductVariantResource extends AbstractResource
@@ -26,6 +27,7 @@ class ProductVariantResource extends AbstractResource
             'unit_qty' => $this->unit_qty,
             'min_qty' => $this->min_qty,
             'max_qty' => $this->max_qty,
+            'min_batch' => $this->min_batch,
             'inventory' => $this->stock,
             'incoming' => $this->incoming,
             'group_pricing' => (bool) $this->group_pricing,
@@ -57,6 +59,7 @@ class ProductVariantResource extends AbstractResource
     {
         return [
             'product' => ['data' => new ProductResource($this->whenLoaded('product'), $this->only)],
+            'image' => new AssetResource($this->whenLoaded('image')),
             'tiers' => new ProductTierCollection($this->whenLoaded('tiers')),
             'tax' => $this->include('tax', TaxResource::class),
         ];